+++ src/main/java/mk/ukim/finki/wp/lab/service/BookReservationServiceImpl.java
@@ -1,14 +1,12 @@
 package mk.ukim.finki.wp.lab.service;
 
-import mk.ukim.finki.wp.lab.bootstrap.DataHolder;
-import mk.ukim.finki.wp.lab.model.Book;
 import mk.ukim.finki.wp.lab.model.BookReservation;
 import mk.ukim.finki.wp.lab.repository.BookReservationRepository;
+import mk.ukim.finki.wp.lab.service.BookReservationService;
 import org.springframework.stereotype.Service;
 
 @Service
 public class BookReservationServiceImpl implements BookReservationService{
-
     private final BookReservationRepository bookReservationRepository;
 
     public BookReservationServiceImpl(BookReservationRepository bookReservationRepository) {
@@ -17,7 +15,7 @@ public class BookReservationServiceImpl implements BookReservationService{
 
     @Override
     public BookReservation placeReservation(String bookTitle, String readerName, String readerAddress, int numberOfCopies) {
-        BookReservation r= new BookReservation(bookTitle,readerName,readerAddress,(long)numberOfCopies);
+        BookReservation r= new BookReservation(bookTitle,readerName,readerAddress, numberOfCopies);
         return  bookReservationRepository.save(r);
     }
 }
